requirement.txt
request==2.22.0
scikit-learn==0.21.3
google-cloud-storage==1.17.0
flask==1.0.2
